+++ library/Zend/Http/Header/SetCookie.php
@@ -302,7 +302,6 @@ class SetCookie implements MultipleHeaderInterface
      */
     public function setValue($value)
     {
-        HeaderValue::assertValid($value);
         $this->value = $value;
         return $this;
     }